% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nested.R
\name{extract_nested}
\alias{extract_nested}
\title{Extracts values from nested data structure of a XML object.}
\usage{
extract_nested(leaves, size)
}
\arguments{
\item{leaves}{Parsed xml children nodes object.}

\item{self}{string of values "l", "m" or "s".}
}
\value{
A data frame of values extracted from XML file.
}
\description{
Used internally by get_areas_large(), get_areas_middle() and get_areas_small().
}
\seealso{
\code{\link{get_areas_large}}, \code{\link{get_areas_middle}},
\code{\link{get_areas_small}}.
}
\keyword{internal}
