% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_shop_info.R
\name{extract_shop_info}
\alias{extract_shop_info}
\title{Extracts shop info from nested data structure of a XML object.}
\usage{
extract_shop_info(leaves, lang = "en")
}
\arguments{
\item{leaves}{Parsed xml children nodes object.}

\item{lang}{string, language of the returned data. Default = "en", can 
also take values of "ja", "zh_cn", "zh_tw", "ko".}
}
\value{
A data frame of shop info extracted from XML file.
}
\description{
Used internally by get_shops().
}
\seealso{
\code{\link{get_shops}}.
}
