% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas.R
\name{get_areas}
\alias{get_areas}
\title{Implements the Acquire Area (regions broader than prefectures) Master API.}
\usage{
get_areas(api_key = "eca7388c8a3c6332eb702a21bcc63b46", lang = "en")
}
\arguments{
\item{api_key}{string, your own Gurunavi API key. Ours has been provided
for convenience.}

\item{lang}{string, language of the returned data. Default = "en", can
also take values of "ja", "zh_cn", "zh_tw", "ko".}
}
\value{
A data frame of area codes and names.
}
\description{
\url{http://api.gnavi.co.jp/api/manual_e.html#api02}
}
\examples{
get_areas(lang="en")
get_areas(lang="ja")
get_areas(lang="zh_cn")
get_areas(lang="zh_tw")
get_areas(lang="ko")
}
\seealso{
\code{\link{query_data}}, \code{\link{extract_simple}}
}
