% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas_middle.R
\name{get_areas_middle}
\alias{get_areas_middle}
\title{Implements the Acquire Area Middle Master API.}
\usage{
get_areas_middle(api_key = "eca7388c8a3c6332eb702a21bcc63b46", lang = "en")
}
\arguments{
\item{api_key}{string, your own Gurunavi API key. Ours has been provided
for convenience.}

\item{lang}{string, language of the returned data. Default = "en", can
also take values of "ja", "zh_cn", "zh_tw", "ko".}
}
\value{
A data frame of middle and large area codes and names and pref codes and names.
}
\description{
\url{http://api.gnavi.co.jp/api/manual_e.html#api05}
}
\examples{
get_areas_middle(lang = "en") 
get_areas_middle(lang = "ja")
get_areas_middle(lang = "zh_cn")
get_areas_middle(lang = "zh_tw") 
get_areas_middle(lang = "ko")
}
\seealso{
\code{\link{query_data}}, \code{\link{extract_nested}}
}
