% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bizcats.R
\name{get_bizcats}
\alias{get_bizcats}
\title{Implements the Acquire Business Category Master API.}
\usage{
get_bizcats(api_key = "eca7388c8a3c6332eb702a21bcc63b46", lang = "en")
}
\arguments{
\item{api_key}{string, your own Gurunavi API key. Ours has been provided
for convenience.}

\item{lang}{string, language of the returned data. Default = "en", can
also take values of "ja", "zh_cn", "zh_tw", "ko".}
}
\value{
A data frame of biz category codes and names.
}
\description{
\url{http://api.gnavi.co.jp/api/manual_e.html#api07}
}
\examples{
get_bizcats(lang = "en")
get_bizcats(lang = "ja")
get_bizcats(lang = "zh_cn")
get_bizcats(lang = "zh_tw")
get_bizcats(lang = "ko")
}
\seealso{
\code{\link{query_data}}, \code{\link{extract_simple}}
}
