% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shops.R
\name{get_shops}
\alias{get_shops}
\title{Implements the Multi-language version Restaurant Search API.}
\usage{
get_shops(api_key = "eca7388c8a3c6332eb702a21bcc63b46", lang = "en",
  pref = "", areacode_s = "", category_l = "", category_s = "",
  english_speaking = 0, korean_speaking = 0, chinese_speaking = 0,
  english_menu = 0, korean_menu = 0, chinese_menu = 0,
  vegetarian_menu = 0, wifi = 0, card = 0, private_room = 0,
  no_smoking = 1, verbose = TRUE)
}
\arguments{
\item{api_key}{string, your own Gurunavi API key. Ours has been provided
for convenience.}

\item{lang}{string, language of the returned data. Default = "en", can
also take values of "ja", "zh_cn", "zh_tw", "ko".}

\item{pref}{string, prefecture code.}

\item{areacode_s}{string, small area code.}

\item{category_l}{string, main biz category code}

\item{category_s}{string, biz sub-category code}

\item{english_speaking}{0 or 1, where 1 means Yes}

\item{korean_speaking}{0 or 1, where 1 means Yes}

\item{chinese_speaking}{0 or 1, where 1 means Yes}

\item{english_menu}{0 or 1, where 1 means Yes}

\item{korean_menu}{0 or 1, where 1 means Yes}

\item{chinese_menu}{0 or 1, where 1 means Yes}

\item{vegetarian_menu}{0 or 1, where 1 means Yes}

\item{wifi}{0 or 1, where 1 means Yes}

\item{card}{0 or 1, where 1 means Yes}

\item{private_room}{0 or 1, where 1 means Yes}

\item{no_smoking}{0 or 1, where 1 means Yes}

\item{verbose}{TRUE or FALSE, whether to print status messsages}
}
\value{
A data frame of shop info.
}
\description{
\url{http://api.gnavi.co.jp/api/manual_e.html#api01}
}
\examples{
get_shops(lang = "en", pref = "PREF27", areacode_s = "AREAS3144", category_l = "RSFST02000", card = 1)
get_shops(lang = "ja", pref = "PREF27", category_l = "RSFST02000", category_s = "RSFST02001")
get_shops(lang = "zh_cn", pref = "PREF27", chinese_menu = 1, chinese_speaking = 1)
get_shops(lang = "zh_tw", pref = "PREF27", chinese_menu = 1, wifi = 1, card = 1)
get_shops(lang = "ko", pref = "PREF27", korean_speaking = 1, korean_menu = 1)
}
\seealso{
\code{\link{query_data}}, \code{\link{extract_shop_info}}
}
