% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_APIcall.R
\name{make_APIcall}
\alias{make_APIcall}
\title{Extracts values from simple flat data structure of a XML object.}
\usage{
make_APIcall(base_url, params)
}
\arguments{
\item{base_url}{string, base url of API.}

\item{params}{list of parameter values.}
}
\value{
url of full API call.
}
\description{
Used internally by query_data().
}
\seealso{
\code{\link{query_data}}.
}
\keyword{internal}
