% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{Performs GET call to API and parses returned XML.}
\usage{
query_data(base_url, params)
}
\arguments{
\item{base_url}{string, base url of API.}

\item{params}{list of search parameters that will make up the URL of a full API call.}
}
\value{
A parsed xml children object.
}
\description{
Used internally by all the exported functions.
}
